+++ package.json
@@ -2,27 +2,20 @@
   "name": "music-bot",
   "main": "main.js",
   "author": "ShambaC",
-  "version": "30.0.0",
+  "version": "2.0.0",
   "repository": "https://github.com/ShambaC/Discord-Music-BOT-Shamba.git",
   "description": "A complete code to download for a cool Discord music bot.",
   "scripts": {
     "start": "node main.js"
   },
   "dependencies": {
-    "@discord-player/extractor": "^4.1.2",
+    "@discord-player/extractor": "^4.5.1",
     "@discordjs/opus": "^0.9.0",
-    "@discordjs/voice": "^0.16.0",
-    "@distube/ytdl-core": "^4.11.8",
-    "discord-player": "^6.1.1",
-    "discord.js": "^14.9.0",
-    "dotenv": "^16.0.2",
-    "genius-lyrics": "^4.3.0",
-    "ms": "^3.0.0-canary.1",
-    "opusscript": "^0.0.8",
-    "prism-media": "^1.3.5",
-    "ytdl-core": "^4.11.3"
-  },
-  "engines": {
-    "node": "16.x"
+    "@evan/opus": "^1.0.3",
+    "discord-player": "^6.7.1",
+    "discord-player-youtubei": "^1.3.0",
+    "discord.js": "^14.15.3",
+    "dotenv": "^16.4.5",
+    "ms": "^2.1.3"
   }
 }
